+++ src/main/java/lotto/domain/Prize.java
@@ -1,7 +1,6 @@
 package lotto.domain;
 
 import java.util.Arrays;
-import java.util.LinkedHashMap;
 
 public enum Prize {
 
@@ -30,17 +29,4 @@ public enum Prize {
                 .orElse(0);
     }
 
-    public static LinkedHashMap<Prize, Integer> calculateStatistic(int totalPrize) {
-        LinkedHashMap<Prize, Integer> statistic = new LinkedHashMap<>();
-        int prize = totalPrize;
-        for (Prize price : Prize.values()) {
-            statistic.put(price, prize / price.getValue());
-            prize %= price.getValue();
-        }
-        return statistic;
-    }
-
-    public static double calculateProfit(int prize, int buyPrice) {
-        return Math.floor(prize / (double) buyPrice * 100) / 100.0;
-    }
 }
